Refactor zonemd verify code out of pdnsutil
